Updated ios build. Renamed bullet folder. Changed the bone shaders to expect floats instead of ints for influence bone index. Various xcode warning fixes. Override isreadytorender, set to always true on ios.

+++ GameEntityComponentTest/Game/Data/Shaders/Include/Bone_AttribsAndUniforms.glsl
@@ -1,7 +1,7 @@
 #if NUM_INF_BONES > 0
     const int MAX_BONES = 100;
 
-    attribute ivec4 a_BoneIndex;
+    attribute vec4 a_BoneIndex;
     attribute vec4 a_BoneWeight;
 
     uniform mat4 u_BoneTransforms[MAX_BONES];
